% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_server_mkplot.R
\name{mod_server_mkplot}
\alias{mod_server_mkplot}
\title{Plotting Server}
\usage{
mod_server_mkplot(id, simdat_raw)
}
\arguments{
\item{id}{Namespace}

\item{simdat_raw}{simulation data.frame}
}
\value{
Shiny server module
}
\description{
Plotting Server
}
