% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_server_runsim.R
\name{mod_server_runsim}
\alias{mod_server_runsim}
\title{Simulation Server}
\usage{
mod_server_runsim(id, dosedat)
}
\arguments{
\item{id}{Namespace}

\item{dosedat}{dosing data.frame}
}
\value{
Shiny server module
}
\description{
Simulation Server
}
